requirements.txt
flask
google-generativeai
requests
dotenv